+++ test/tx-data/vote-commission.test.js
@@ -50,6 +50,9 @@ describe('TxDataVoteCommission', () => {
         voteCommission: 12,
         voteUpdate: 12,
         createSwapPool: 12,
+        failedTx: 12,
+        addLimitOrder: 12,
+        removeLimitOrder: 12,
     };
 
     const serializedTxData = (new TxDataVoteCommission(params)).serialize();
@@ -62,6 +65,9 @@ describe('TxDataVoteCommission', () => {
                 [],
                 // [12],
                 // [12],
+                [12],
+                [12],
+                [12],
                 [12],
                 [12],
                 [12],